requirements.txt
Flask==2.0.2
python-docx==0.8.11
pandas==1.3.3
fuzzywuzzy==0.18.0
openpyxl